pyproject.toml
[tool.poetry]
name = "paraicehockey"
version = "0.1.0"
description = ""
authors = ["Ekaterina Kolosova <kolosova411@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.15"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.3"
flake8-docstrings = "^1.6.0"
flake8-isort = "^5.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
